fix(trusty-mpm): drop decommissioned sessions from TUI Sessions pane
Decommissioning a session outside the `tm projects` TUI flips its state to
"decommissioned" in the daemon's session store but never deletes the
record, so the unfiltered fleet->row projection let a dead session's row
persist in the Sessions pane forever and inflated its (N) header count.

Fix: filter decommissioned sessions out of the row projection
(live_session_rows, new in poll/rows.rs) before they ever reach
sessions_by_project, so the pane and its header count reflect only live
sessions. Since sessions_by_project is rebuilt wholesale from the daemon's
fleet snapshot on every poll tick, a session later replaced by a new one
reusing its name can never ghost or duplicate.

Split poll.rs into poll/mod.rs (async polling) + poll/rows.rs (pure DTO ->
row projections) to stay under the 500-SLOC production cap after the new
function and its tests. Updated DOC-35 §5.1 to document the resolved
behavior.

Closes #2476
